requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.9.6
aniso8601==0.82
facebook-sdk==0.4.0
itsdangerous==0.24
peewee==2.2.5
pytz==2014.4
six==1.7.3
wsgiref==0.1.2
